move all runtime manager code into one file
